third_party/blink/web_tests/wpt_internal/fenced_frame/csp-urn.https.html

<!DOCTYPE html>
<title>Test opaque fenced frame interacting with CSP</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
function setupCSP(csp, second_csp=null) {
  let meta = document.createElement('meta');
  meta.httpEquiv = "Content-Security-Policy";
  meta.content = "fenced-frame-src " + csp;
  document.head.appendChild(meta);

  if (second_csp != null) {
    let second_meta = document.createElement('meta');
    second_meta.httpEquiv = "Content-Security-Policy";
    second_meta.content = "frame-src " + second_csp;
    document.head.appendChild(second_meta);
  }
}

const allowedCSPs = ["*", "https:", "https://*:*"];
allowedCSPs.forEach((csp) => {
  promise_test(async() => {
    setupCSP(csp);

    const key = token();

    window.addEventListener('securitypolicyviolation', function(e) {
      // Write to the server even though the listener is in the same file in the test below.
      writeValueToServer(key, e.violatedDirective + ";" + e.blockedURI);
    });

    attachFencedFrame(await generateURN("resources/embeddee.html", [key]), "opaque-ads");

    const result = await nextValueFromServer(key);
    assert_equals(result, "PASS",
        "The fenced frame should load for CSP fenced-frame-src " + csp);
    }, "Fenced frame loaded for CSP fenced-frame-src " + csp);

  promise_test(async() => {
    setupCSP(csp);
    assert_true(HTMLFencedFrameElement.canLoadOpaqueURL());
  }, "Opaque-ads can load API returns true for " + csp);
});

const blockedCSPs = ["'none'", "'self'", "data:", "https://*", "https://*:80", "https://b.test:*"];
blockedCSPs.forEach((csp) => {
  promise_test(async() => {
    setupCSP(csp);

    const key = token();

    window.addEventListener('securitypolicyviolation', function(e) {
      // Write to the server even though the listener is in the same file in the test below.
      writeValueToServer(key, e.violatedDirective + ";" + e.blockedURI);
    });

    attachFencedFrame(await generateURN("resources/embeddee.html", [key]), "opaque-ads");

    const result = await nextValueFromServer(key);
    assert_equals(result, "fenced-frame-src;",
        "The fenced frame should not load for CSP fenced-frame-src " + csp);
  }, "Fenced frame blocked for CSP fenced-frame-src " + csp);

  promise_test(async() => {
    setupCSP(csp);
    assert_false(HTMLFencedFrameElement.canLoadOpaqueURL());
  }, "Opaque-ads can load API returns false for " + csp);
});

promise_test(async() => {
    setupCSP("*", "'self'");

    const key = token();

    window.addEventListener('securitypolicyviolation', function(e) {
      // Write to the server even though the listener is in the same file in the test below.
      writeValueToServer(key, e.violatedDirective + ";" + e.blockedURI);
    });

    attachFencedFrame(await generateURN("resources/embeddee.html", [key]), "opaque-ads");

    const result = await nextValueFromServer(key);
    assert_equals(result, "fenced-frame-src;",
        "The fenced frame should not load for CSP frame-src 'self' even if " +
        "another CSP allows loading a fenced frame.");

    // Test the canLoadOpaqueURL API to ensure it arrives at the same result.
    assert_false(HTMLFencedFrameElement.canLoadOpaqueURL());
}, "Fenced frame not loaded if any of CSPs in place disallow loading");
</script>
</body>
